block refactoring
