+++ exonum-java-binding/common/src/test/java/com/exonum/binding/common/blockchain/BlockTest.java
@@ -1,5 +1,5 @@
 /*
- * Copyright 2018 The Exonum Team
+ * Copyright 2019 The Exonum Team
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package com.exonum.binding.blockchain;
+package com.exonum.binding.common.blockchain;
 
 import com.exonum.binding.common.hash.HashCode;
 import nl.jqno.equalsverifier.EqualsVerifier;
