+++ exonum-java-binding/core/src/main/java/com/exonum/binding/blockchain/CoreSchemaProxy.java
@@ -23,13 +23,14 @@ import static com.google.common.base.Preconditions.checkArgument;
 import com.exonum.binding.blockchain.serialization.BlockSerializer;
 import com.exonum.binding.blockchain.serialization.TransactionLocationSerializer;
 import com.exonum.binding.blockchain.serialization.TransactionResultSerializer;
+import com.exonum.binding.common.blockchain.Block;
+import com.exonum.binding.common.blockchain.TransactionLocation;
+import com.exonum.binding.common.blockchain.TransactionResult;
 import com.exonum.binding.common.configuration.StoredConfiguration;
 import com.exonum.binding.common.hash.HashCode;
 import com.exonum.binding.common.message.TransactionMessage;
 import com.exonum.binding.common.serialization.Serializer;
 import com.exonum.binding.common.serialization.StandardSerializers;
-import com.exonum.binding.common.blockchain.TransactionLocation;
-import com.exonum.binding.common.blockchain.TransactionResult;
 import com.exonum.binding.proxy.Cleaner;
 import com.exonum.binding.proxy.NativeHandle;
 import com.exonum.binding.proxy.ProxyDestructor;
