+++ exonum-java-binding/core/src/test/java/com/exonum/binding/blockchain/BlockchainTest.java
@@ -21,10 +21,11 @@ import static org.assertj.core.api.Assertions.assertThat;
 import static org.mockito.Mockito.mock;
 import static org.mockito.Mockito.when;
 
-import com.exonum.binding.common.configuration.StoredConfiguration;
-import com.exonum.binding.common.hash.HashCode;
+import com.exonum.binding.common.blockchain.Block;
 import com.exonum.binding.common.blockchain.TransactionLocation;
 import com.exonum.binding.common.blockchain.TransactionResult;
+import com.exonum.binding.common.configuration.StoredConfiguration;
+import com.exonum.binding.common.hash.HashCode;
 import com.exonum.binding.storage.indices.ListIndexProxy;
 import com.exonum.binding.storage.indices.MapIndex;
 import com.exonum.binding.storage.indices.ProofListIndexProxy;
