+++ exonum-java-binding/qa-service/src/main/java/com/exonum/binding/qaservice/QaService.java
@@ -16,12 +16,12 @@
 
 package com.exonum.binding.qaservice;
 
-import com.exonum.binding.blockchain.Block;
+import com.exonum.binding.common.blockchain.Block;
+import com.exonum.binding.common.blockchain.TransactionLocation;
+import com.exonum.binding.common.blockchain.TransactionResult;
 import com.exonum.binding.common.configuration.StoredConfiguration;
 import com.exonum.binding.common.hash.HashCode;
 import com.exonum.binding.common.message.TransactionMessage;
-import com.exonum.binding.common.blockchain.TransactionLocation;
-import com.exonum.binding.common.blockchain.TransactionResult;
 import com.exonum.binding.service.Service;
 import java.util.List;
 import java.util.Map;
